Moved and separated tests to fit folder structure
