+++ tests/Sharpify.Tests/Routines/AsyncRoutineTests.cs
@@ -1,26 +1,8 @@
 using Sharpify.Routines;
 
-namespace Sharpify.Tests;
+namespace Sharpify.Tests.Routines;
 
-public class RoutineTests {
-    [Theory]
-    [InlineData(550, 5)]
-    [InlineData(150, 1)]
-    [InlineData(1050, 10)]
-    [InlineData(50, 0)]
-    public async Task Routine_GivenIncreaseValueFunction_IncreasesValue(int milliseconds, int expected) {
-        // Arrange
-        var count = 0;
-        using var routine = new Routine(100).Add(() => count++);
-
-        // Act
-        routine.Start();
-        await Task.Delay(milliseconds);
-
-        // Assert
-        count.Should().Be(expected);
-    }
-
+public class AsyncRoutineTests {
     [Theory]
     [InlineData(550, 5)]
     [InlineData(150, 1)]
